requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.5
chromadb==0.4.18
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
numpy==1.24.3
beautifulsoup4==4.12.2